requirements.txt
coloredlogs~=15.0
discord.py~=1.6
